pyproject.toml
[project]
name = "mx-smc-pa-trading"
version = "2.0.0"
description = "MX SMC+PA quantitative trading system with live execution & backtesting"
requires-python = ">=3.10"
dependencies = [
    # Live trading
    "fastapi>=0.115.0,<1.0",
    "uvicorn>=0.32.0,<1.0",
    "pydantic>=2.9.0,<3.0",
    "pydantic-settings>=2.0.0,<3.0",
    "python-dotenv>=1.0.1,<2.0",
    "loguru>=0.7.3,<1.0",
    "pyyaml>=6.0,<7.0",

    # Brokers
    "shioaji>=1.5.0,<2.0",
    "sj-sync>=0.1.0",
    "async-rithmic>=1.0.0,<2.0",

    # Backtesting
    "backtrader>=1.9.78,<2.0",
    "optuna>=3.0.0,<4.0",
    "pandas>=2.0.0,<3.0",
    "numpy>=1.24.0,<2.0",

    # Reporting
    "jinja2>=3.1.0,<4.0",
    "plotly>=5.0.0,<6.0",
    "matplotlib>=3.7.0,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",          # for FastAPI TestClient
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
